# Conflicts:
#	src/java.base/share/classes/java/lang/Integer.java
#	src/java.base/share/classes/java/lang/Long.java

+++ test/hotspot/jtreg/runtime/FieldLayout/BaseOffsets.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022, 2024, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2022, 2025, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -110,8 +110,8 @@ public class BaseOffsets {
     public static final WhiteBox WB = WhiteBox.getWhiteBox();
 
     static final long INT_OFFSET;
-    static final int  INT_ARRAY_OFFSET;
-    static final int  LONG_ARRAY_OFFSET;
+    static final long INT_ARRAY_OFFSET;
+    static final long LONG_ARRAY_OFFSET;
     static {
         if (!Platform.is64bit() || WB.getBooleanVMFlag("UseCompactObjectHeaders")) {
             INT_OFFSET = 8;
@@ -151,7 +151,7 @@ public class BaseOffsets {
         Asserts.assertEquals(unsafe.arrayBaseOffset(double[].class),  LONG_ARRAY_OFFSET, "Misplaced double  array base");
         boolean narrowOops = System.getProperty("java.vm.compressedOopsMode") != null ||
                              !Platform.is64bit();
-        int expected_objary_offset = narrowOops ? INT_ARRAY_OFFSET : LONG_ARRAY_OFFSET;
+        long expected_objary_offset = narrowOops ? INT_ARRAY_OFFSET : LONG_ARRAY_OFFSET;
         Asserts.assertEquals(unsafe.arrayBaseOffset(Object[].class),  expected_objary_offset, "Misplaced object  array base");
     }
 }
